Move FHT to dedicated file
